requirements.txt
feedparser==6.0.11
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
pytest==8.4.2

